Test person and image done

+++ src/core/image/infrastructure/aws/image.aws.repository.ts
@@ -5,17 +5,21 @@ import storage from '../../../../sdks/aws.sdk'
 import { PutObjectCommand } from '@aws-sdk/client-s3'
 
 export default class ImageAwsRepository implements UploaderRepository {
-  async uploadImage (path: string, fileName: string) {
+  async uploadImage (path: string, fileName: string, isUnlinkeable?: boolean): Promise<string> {
     try {
       await storage.send(new PutObjectCommand({
         Bucket: process.env.AWS_BUCKET_NAME,
         Key: fileName,
         Body: fsNonPromise.createReadStream(path)
       }))
-      await fs.unlink(path)
+      if (isUnlinkeable) {
+        await fs.unlink(path)
+      }
       return `https://${process.env.AWS_BUCKET_NAME}.s3.${process.env.AWS_REGION}.amazonaws.com/${fileName}`
     } catch (error) {
-      await fs.unlink(path)
+      if (isUnlinkeable) {
+        await fs.unlink(path)
+      }
       return ''
     }
   }
