+++ src/core/image/infrastructure/cloudinary/image.cloudinary.respository.ts
@@ -3,14 +3,17 @@ import cloudinary from '../../../../sdks/cloudinary.sdk'
 import fs from 'node:fs/promises'
 
 export default class ImageCloudinaryRepository implements UploaderRepository {
-  async uploadImage (path: string) {
+  async uploadImage (path: string, fileName: string, isUnlinkeable?: boolean): Promise<string> {
     try {
       const { secure_url: secureUrl } = await cloudinary.uploader.upload(path)
-      await fs.unlink(path)
+      if (isUnlinkeable) {
+        await fs.unlink(path)
+      }
       return secureUrl
     } catch (error) {
-      console.log(error)
-      await fs.unlink(path)
+      if (isUnlinkeable) {
+        await fs.unlink(path)
+      }
       return ''
     }
   }
